seperating entity and battle test & setup beforeach for printer + adding a test to see who is the winner
