+++ src/EntityTest.java
@@ -3,17 +3,11 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.function.Executable;
-
-
 
 import static org.junit.jupiter.api.Assertions.*;
 
-// FIXME: remove this class
-public class DummyTest {
-
-       
+public class EntityTest {
     @Test
     @DisplayName("Test in Entity")
     public void testInConstructor() {
@@ -44,7 +38,6 @@ public class DummyTest {
         // Assert
         assertEquals(0, entity.getHealth(), "Wrong health");
     }
-
 
     @Test
     @DisplayName("Test if entity is dead")
@@ -85,7 +78,6 @@ public class DummyTest {
         assertEquals(2, entity.getCooldown(), "Wrong cooldown");
     }
 
-   
     @Test
     @DisplayName("Test attack value")
     public void testAttackBiggerThanZero() {
@@ -112,8 +104,6 @@ public class DummyTest {
         // Assert
         assertEquals(defense, entity.getDefense(), "Wrong health");
     }
-
-
 
     @Test
     @DisplayName("Test attack with attack() function")
@@ -142,79 +132,5 @@ public class DummyTest {
 
         // Assert
         assertThrows(IllegalArgumentException.class, executable, "Exception handling failed");
-    }
-
-    @Test
-    @DisplayName("Test for battle function")
-    public void testForBattle1() {
-        // Arrange
-        Entity entity1 = new Entity(100, 50, 35, 2, "Bob", "Bob story");
-        Entity entity2 = new Entity(90, 40, 30, 2, "Mike", "Mike story");
-        var printer = System.out;
-        Battle battle = new Battle(printer);
-
-        // //Act
-        battle.battle(entity1, entity2);
-
-        // Assert
-        assertTrue(entity1.getHealth() > entity2.getHealth(), "Wrong health");
-    }
-
-    @ParameterizedTest
-    @ValueSource(doubles = {90, 50, 70, 80, 40, 45.7, 27.9})
-    @DisplayName("Test for battle function 2")
-    public void testForBattle2(double health) {
-        // Arrange
-        Entity entity1 = new Entity(100, 50, 35, 2, "Bob", "Bob story");
-        Entity entity2 = new Entity(health, 40, 30, 2, "Mike", "Mike story");
-        var printer = System.out;
-        Battle battle = new Battle(printer);
-
-        //Act
-        battle.battle(entity1, entity2);
-
-        // Assert
-        assertTrue(entity1.isAlive(), "Wrong health");
-    }
-
-    @ParameterizedTest
-    @ValueSource(doubles = {17, 30, 50, 40, 15, 45.7, 27.9})
-    @DisplayName("Test for battle function 2")
-    public void testForBattle3(double attack) {
-        // Arrange
-        Entity entity1 = new Entity(100, 50, 35, 2, "Bob", "Bob story");
-        Entity entity2 = new Entity(90, attack, 30, 2, "Mike", "Mike story");
-        var printer = System.out;
-        Battle battle = new Battle(printer);
-
-        //Act
-        battle.battle(entity1, entity2);
-
-        // Assert
-        assertTrue(entity1.getHealth() > entity2.getHealth(), "Wrong health");
-    }
-
-    @ParameterizedTest
-    @ValueSource(doubles = {17, 30, 20, 30, 15, 15.7, 27.9})
-    @DisplayName("Test for battle function 2")
-    public void testForBattle4(double defense) {
-        // Arrange
-        Entity entity1 = new Entity(100, 50, 35, 2, "Bob", "Bob story");
-        Entity entity2 = new Entity(90, 40, defense, 2, "Mike", "Mike story");
-        var printer = System.out;
-        Battle battle = new Battle(printer);
-
-
-        //Act
-        battle.battle(entity1, entity2);
-
-        // Assert
-        assertTrue(entity1.isAlive(), "Wrong Parameters");
-    }
-  
-    
-    @Test
-    public void dummyTest() {
-        assertTrue(true);
     }
 }
